% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_anot.R
\name{anot_susie}
\alias{anot_susie}
\title{Single effect Bayesian regression model}
\usage{
anot_susie(
  y,
  X,
  L,
  pi = NULL,
  credible_prob = 0.95,
  sigma = 1,
  sigma_0 = 1,
  z_cov
)
}
\arguments{
\item{y}{vector of phenotypes standardized and centered.}

\item{X}{SNP matrix of dimenstion n by p (centered and standardized)}

\item{L}{Number of credible set}

\item{pi}{vector of length p represent prior probability of the SNPs}

\item{credible_prob}{probability of the credible set containing atleast 1 causal SNP}

\item{sigma}{prior sigma}

\item{sigma_0}{prior probablity for beta}

\item{z_cov}{matrix of annotation}
}
\description{
This function allows you to express your love of cats.
}
\examples{
anot_susie()
}
