% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsimp.R
\name{Bayes_simp}
\alias{Bayes_simp}
\title{Simple Bayesian linear regression model}
\usage{
Bayes_simp(y, x, sigma, sigma_0)
}
\arguments{
\item{y}{vector of phenotypes standardized and centered.}

\item{sigma}{prior sigma}

\item{sigma_0}{prior probablity for beta}

\item{X}{SNP matrix of dimenstion n by p (centered and standardized)}
}
\description{
This function allows you to express your love of cats.
}
\examples{
Bayes_simp()
}
