% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibss_fun.R
\name{IBSS_function}
\alias{IBSS_function}
\title{IBSS function}
\usage{
IBSS_function(y, X, L, pi = NULL, credible_prob = 0.95, sigma = 1, sigma_0 = 1)
}
\arguments{
\item{y}{vector of phenotypes standardized and centered.}

\item{X}{SNP matrix of dimenstion n by p (centered and standardized)}

\item{pi}{vector of length p represent prior probability of the SNPs}

\item{credible_prob}{probability of the credible set containing atleast 1 causal SNP}

\item{sigma}{prior sigma}

\item{sigma_0}{prior probablity for beta}
}
\description{
This function allows you to express your love of cats.
}
\examples{
IBSS_function()
}
